Adapt the gem's structure to fit better with current standards
