+++ Gemfile
@@ -1,3 +1,6 @@
+# frozen_string_literal: true
+
 source 'https://rubygems.org'
-gem 'rspec'
-gem 'msgpack'
+
+# Specify your gem's dependencies in markov_noodles.gemspec
+gemspec
