+++ bin/console
@@ -1,0 +1,8 @@
+#!/usr/bin/env ruby
+# frozen_string_literal: true
+
+require 'bundler/setup'
+require 'markov_noodles'
+
+require 'irb'
+IRB.start(__FILE__)
