+++ lib/markov_noodles/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class MarkovNoodles
+  VERSION = '1.0.0'
+end
